Moves SubmissionConversion to own app
