+++ app/tests/evaluation_tests/test_tasks.py
@@ -5,7 +5,7 @@ import pytest
 from django.core.exceptions import ValidationError
 
 from grandchallenge.container_exec.tasks import validate_docker_image_async
-from grandchallenge.datasets.models import SubmissionConversionJob
+from grandchallenge.submission_conversion.models import SubmissionConversionJob
 from grandchallenge.evaluation.models import Method
 from tests.factories import SubmissionFactory, MethodFactory, ImageSetFactory
